+++ src/main/java/mk/edu/uklo/fikt/fiktexamweb/util/QuestionRepository.java
@@ -10,6 +10,5 @@ import java.util.List;
 
 @Repository
 public interface QuestionRepository extends JpaRepository<Question, Integer>{
-    List<Question> findByTopicId(long subjectId);
-
+    List<Question> findByTopicId(int subjectId);
 }
